+++ python/lsst/ts/salobj/topics/write_topic.py
@@ -21,10 +21,12 @@ from __future__ import annotations
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-__all__ = ["MAX_SEQ_NUM", "WriteTopic"]
+__all__ = ["MAX_SEQ_NUM", "SetWriteResult", "WriteTopic"]
 
+import copy
 import struct
 import typing
+import warnings
 
 import numpy as np
 
@@ -45,8 +47,31 @@ if typing.TYPE_CHECKING:
 MAX_SEQ_NUM = (1 << 31) - 1
 
 
+class SetWriteResult:
+    """Result from set_write.
+
+    Parameters
+    ----------
+    did_change
+        Did the keywords change any values?
+    was_written
+        Was the message written?
+    data
+        The data, after applying keywords.
+        This should be a copy, to avoid race conditions between this
+        call and other code that alters the data.
+    """
+
+    def __init__(
+        self, did_change: bool, was_written: bool, data: type_hints.BaseDdsDataType
+    ) -> None:
+        self.did_change = did_change
+        self.was_written = was_written
+        self.data = data
+
+
 class WriteTopic(BaseTopic):
-    r"""Base class for topics that are output.
+    r"""Base class for topics that are written.
 
     This includes  controller events, controller telemetry, remote commands
     and ``cmdack`` writers.
@@ -76,6 +101,8 @@ class WriteTopic(BaseTopic):
     isopen : `bool`
         Is this instance open? `True` until `close` or `basic_close` is called.
     """
+    # Default value for the force_output argument of write
+    default_force_output = True
 
     def __init__(
         self,
@@ -157,7 +184,11 @@ class WriteTopic(BaseTopic):
 
     @property
     def has_data(self) -> bool:
-        """Has `data` ever been set?"""
+        """Has `data` ever been set?
+
+        Note: a value of true means at least one field has been set,
+        not that all fields have been set.
+        """
         return self._has_data
 
     def basic_close(self) -> None:
@@ -178,11 +209,8 @@ class WriteTopic(BaseTopic):
         """
         self.basic_close()
 
-    def put(
-        self,
-        data: typing.Optional[type_hints.BaseDdsDataType] = None,
-    ) -> None:
-        """Output this topic.
+    def put(self, data: typing.Optional[type_hints.BaseDdsDataType] = None) -> None:
+        """DEPRECATED: call async method `write` instead. Write this topic.
 
         Parameters
         ----------
@@ -195,37 +223,26 @@ class WriteTopic(BaseTopic):
             If ``data`` is not None and not an instance of `DataType`.
         """
         if data is not None:
+            warnings.warn(
+                "put with data is deprecated; call write with kwargs instead",
+                DeprecationWarning,
+            )
             self.data = data
+        else:
+            warnings.warn("put is deprecated; call write instead", DeprecationWarning)
 
-        self.data.private_sndStamp = utils.current_tai()
-        self.data.private_revCode = self.rev_code
-        self.data.private_origin = self.salinfo.domain.origin
-        self.data.private_identity = self.salinfo.identity
-        if self._seq_num_generator is not None:
-            self.data.private_seqNum = next(self._seq_num_generator)
-        # when index is 0 use the default of 0 and give senders a chance
-        # to override it.
-        if self.salinfo.index != 0:
-            setattr(self.data, f"{self.salinfo.name}ID", self.salinfo.index)
-        try:
-            self._writer.write(self.data)
-        except struct.error as e:
-            raise ValueError(
-                f"{self.name} write({self.data}) failed: one or more fields invalid"
-            ) from e
-        except TypeError as e:
-            raise ValueError(
-                f"{self.name} write({self.data}) failed: "
-                f"perhaps one or more array fields has been set to a scalar"
-            ) from e
-        except IndexError as e:
-            raise ValueError(
-                f"{self.name} write({self.data}) failed: "
-                f"probably one or more array fields is too short"
-            ) from e
+        self._basic_write()
 
     def set(self, **kwargs: typing.Any) -> bool:
-        """Set one or more fields of message data cache ``self.data``.
+        """Set one or more fields of message data ``self.data``.
+
+        This is useful when accumulating data for a topic in different
+        bits of code. Have each bit of code call `set` to set the fields
+        it knows about. Have the last bit of code call `write` to set
+        the remaining fields and write the completed message.
+
+        If you have all the information for a topic in one place, it is simpler
+        to call `write` to set all of the fields and write the message.
 
         Parameters
         ----------
@@ -280,3 +297,85 @@ class WriteTopic(BaseTopic):
                 ) from e
         self._has_data = True
         return did_change
+
+    async def set_write(
+        self, *, force_output: typing.Optional[bool] = None, **kwargs: typing.Any
+    ) -> SetWriteResult:
+        """Set zero or more fields of ``self.data`` and write if changed
+        or if ``force_output`` true.
+
+        Parameters
+        ----------
+        force_output : `bool`, optional
+            If True then write the event, even if no fields have changed.
+            If None (the default), use the class default,
+            which is True for all except ControllerEvent.
+            (The default value is given by class constant
+            ``default_force_output``).
+        **kwargs : `dict` [`str`, ``any``]
+            The remaining keyword arguments are
+            field name = new value for that field.
+            See `set` for more information about values.
+
+        Returns
+        -------
+        result
+            The resulting data and some flags.
+
+        Notes
+        -----
+        The reason there are separate `set_write` and `write` methods is that
+        `write` reliably writes the data, whereas the event version of
+        `set_write` only writes the data if ``kwargs`` changes it,
+        or if ``force_output`` is true.
+        """
+        did_change = self.set(**kwargs)
+        if did_change:
+            do_output = True
+        else:
+            do_output = (
+                self.default_force_output if force_output is None else force_output
+            )
+        if do_output:
+            self._basic_write()
+        return SetWriteResult(
+            did_change=did_change, was_written=do_output, data=copy.copy(self.data)
+        )
+
+    async def write(self) -> type_hints.BaseDdsDataType:
+        """Write the current data and return a copy of the data written.
+
+        Return a copy in order to avoid race conditions.
+        """
+        self._basic_write()
+        return copy.copy(self.data)
+
+    def _basic_write(self) -> None:
+        """Put self.data after setting the private_x fields."""
+        self.data.private_sndStamp = utils.current_tai()
+        self.data.private_revCode = self.rev_code
+        self.data.private_origin = self.salinfo.domain.origin
+        self.data.private_identity = self.salinfo.identity
+        if self._seq_num_generator is not None:
+            self.data.private_seqNum = next(self._seq_num_generator)
+        # when index is 0 use the default of 0 and give senders a chance
+        # to override it.
+        if self.salinfo.index != 0:
+            setattr(self.data, f"{self.salinfo.name}ID", self.salinfo.index)
+
+        try:
+            self._writer.write(self.data)
+        except struct.error as e:
+            raise ValueError(
+                f"{self.name} write({self.data}) failed: one or more fields invalid"
+            ) from e
+        except TypeError as e:
+            raise ValueError(
+                f"{self.name} write({self.data}) failed: "
+                f"perhaps one or more array fields has been set to a scalar"
+            ) from e
+        except IndexError as e:
+            raise ValueError(
+                f"{self.name} write({self.data}) failed: "
+                f"probably at least one array field is too short"
+            ) from e